requirements.txt
Flask
flask-cors
Flask-Limiter
python-dotenv
werkzeug
azure-identity
azure-storage-blob
azure-communication-email
tzdata
clamd
gunicorn
